moved common code into separate file
