+++ src/main/kotlin/org/wfanet/measurement/kingdom/deploy/gcloud/spanner/writers/ClaimReadyExchangeStep.kt
@@ -27,7 +27,6 @@ import org.wfanet.measurement.gcloud.spanner.bufferTo
 import org.wfanet.measurement.gcloud.spanner.insertMutation
 import org.wfanet.measurement.gcloud.spanner.set
 import org.wfanet.measurement.gcloud.spanner.setJson
-import org.wfanet.measurement.gcloud.spanner.updateMutation
 import org.wfanet.measurement.internal.kingdom.ExchangeStep
 import org.wfanet.measurement.internal.kingdom.ExchangeStepAttempt
 import org.wfanet.measurement.internal.kingdom.ExchangeStepAttemptDetails
@@ -137,40 +136,3 @@ class ClaimReadyExchangeStep(
     return row.getLong("MaxAttemptIndex") + 1L
   }
 }
-
-internal fun SpannerWriter.TransactionScope.updateExchangeStepState(
-  exchangeStep: ExchangeStep,
-  recurringExchangeId: Long,
-  state: ExchangeStep.State
-): ExchangeStep {
-  if (exchangeStep.state == state) {
-    return exchangeStep
-  }
-  require(!exchangeStep.state.isTerminal) {
-    "ExchangeStep with StepIndex: ${exchangeStep.stepIndex} is in a terminal state."
-  }
-  val updateTime = Value.COMMIT_TIMESTAMP
-  updateMutation("ExchangeSteps") {
-      set("RecurringExchangeId" to recurringExchangeId)
-      set("Date" to exchangeStep.date.toCloudDate())
-      set("StepIndex" to exchangeStep.stepIndex.toLong())
-      set("State" to state)
-      set("UpdateTime" to updateTime)
-    }
-    .bufferTo(transactionContext)
-
-  return exchangeStep.toBuilder().setState(state).setUpdateTime(updateTime.toProto()).build()
-}
-
-internal val ExchangeStep.State.isTerminal: Boolean
-  get() =
-    when (this) {
-      ExchangeStep.State.BLOCKED,
-      ExchangeStep.State.READY,
-      ExchangeStep.State.READY_FOR_RETRY,
-      ExchangeStep.State.IN_PROGRESS -> false
-      ExchangeStep.State.SUCCEEDED,
-      ExchangeStep.State.FAILED,
-      ExchangeStep.State.UNRECOGNIZED,
-      ExchangeStep.State.STATE_UNSPECIFIED -> true
-    }
