+++ src/test/kotlin/org/wfanet/measurement/kingdom/deploy/gcloud/spanner/SpannerExchangeStepAttemptsServiceTest.kt
@@ -24,8 +24,7 @@ import org.wfanet.measurement.kingdom.deploy.gcloud.spanner.testing.KINGDOM_SCHE
 import org.wfanet.measurement.kingdom.service.internal.testing.ExchangeStepAttemptsServiceTest
 
 @RunWith(JUnit4::class)
-class SpannerExchangeStepAttemptsServiceTest :
-  ExchangeStepAttemptsServiceTest<SpannerExchangeStepAttemptsService>() {
+class SpannerExchangeStepAttemptsServiceTest : ExchangeStepAttemptsServiceTest() {
 
   @get:Rule val spannerDatabase = SpannerEmulatorDatabaseRule(KINGDOM_SCHEMA)
   private val clock = Clock.systemUTC()
